server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
bcrypt==4.1.2
python-jose==3.3.0
python-multipart==0.0.6
motor==3.7.1
pymongo==4.15.4